requirements.txt
aiogram>=3.5.0,<4.0.0
SQLAlchemy>=2.0.29
alembic>=1.13.1
asyncpg>=0.29.0
aiosqlite>=0.19.0
python-dotenv>=1.0.1
redis>=5.0.0
langchain>=0.1.20
langgraph>=0.0.30
langgraph-checkpoint-sqlite>=1.0.0
langgraph-checkpoint-postgres>=1.0.0
langchain-openai>=0.1.6
pydantic>=2.6.0
fastapi>=0.111.0
uvicorn>=0.30.0
django>=5.0.6
dj-database-url>=2.1.0
whitenoise>=6.6.0
openpyxl>=3.1.2
fpdf2>=2.7.6
